+++ mlkem/src/poly.h
@@ -103,7 +103,7 @@ __contract__(
   const uint32_t QINV = 62209;
 
   /*  Compute a*q^{-1} mod 2^16 in unsigned representatives */
-  const uint16_t a_reduced = a & UINT16_MAX;
+  const uint16_t a_reduced = (uint16_t)(a & UINT16_MAX);
   const uint16_t a_inverted = (a_reduced * QINV) & UINT16_MAX;
 
   /* Lift to signed canonical representative mod 2^16. */
